+++ risesin-service-api/src/main/java/com/risesin/service_api/dao/core/PlanChildcaseDao.java
@@ -1,6 +1,6 @@
 package com.risesin.service_api.dao.core;
 
-import com.risesin.service.modules.core.entity.PlanChildcase;
+import com.risesin.service_api.modules.core.entity.PlanChildcase;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
@@ -10,6 +10,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
  * @author Darling
  *
  */
-public interface PlanChildcaseDao extends JpaRepository<PlanChildcase,Integer>,JpaSpecificationExecutor<PlanChildcase>{
+public interface PlanChildcaseDao extends JpaRepository<PlanChildcase,Long>,JpaSpecificationExecutor<PlanChildcase>{
 	
 }
